Print a message on images that cannot be colored instead of the whole chapter failing
try/except around call to colorize_single_image in colorize_images in inference.py

Error messages like "image file is truncated (42 bytes not processed)" were causing the whole batch image process to fail. After this change, the batch process completes and is able to serve the successfully colorized images, but a 404 error occurs when a colored image is requested that could not be generated.