Chrome extension that tries to work on any site
Based on existing Frontend-Chrome.
Using a context menu item instead of embedding a button so we don't need to know anything about the site layout to fit a new button into it.
Instead of using a site-specific class name to find images, try all that are larger than 500x500 pixels.
Instead of getting accurate manga title and chapter from a site-specific place in the DOM, use document.location.hostname for the title and document.location.pathname for the manga name. Not as nice, but it works.
Add permissions:  contextMenus, activeTab, scripting
Remove permissions: host_permissions. content_scripts
Add background service worker to manage context menu and inject contentScript into current tab.
Initial testing finds that this works on the site that already worked with the existing extension but not on a site that requires a captcha.