New backend app-stream uses base64-encoded images
The image to be colored is sent base-64-encoded in the request and the response includes the base-64-encoded colored image.
Instead of writing files to disk, all processing is done in memory.
This means there is no caching in the manga folder.
Previously an entry point to app.py was added to do base-64, but that version also wrote images to disk. Now the base-64 handling has been moved to app-stream.py.
Instead of running the backend as "python app.py" run "python app-stream.py" to use this version.