Check img.complete in colorizeImg
Before trying to colorize, make sure img.complete and img has a name from src or dataset.src.
Safer attempt to get name in case there is no src or dataset.src.
The case if img that is not complete is still not working -
img.addEventListener('load', colorizeImg.bind(img, apiURL));
was supposed to call colorizeImg after img load, but it is still not complete and has no src inside colorizeImg.