Limit requested size, new already-colored method
Reduce maxActiveFetches from 2 to 1 to reduce CUDA out-of-memory errors.
Maximum image width to request from backend, also to reduce CUDA errors.

New maxDistFromGray method being used in place of isColoredContext for testing how far we need to allow from gray.
Color is now allowed to be up to (minDistFromGray = 30) before deciding it must already be a color image that does not need to be colored. (Previously a hard-coded value of 2 was used.) This was added because jpg images tend to have colors farther than expected from pure grays even when they look grayscale.
Plan to add back in early return for colored images after further testing.