Add support for Mangafox (changes img.src to page)
Instead of setting a class on colored images, set coloredsrc attribute to keep track of the most recent image src that was in color.
If image src does not match coloredsrc, need to recolor it.
Only keep the first maxColoredSrc = 200 characters of img.src since recolored src contains all the image data and too many on the same page might stress memory.